requirements.txt
gensim~=4.0
imbalanced-learn~=0.8
jupyterlab
keras
matplotlib
nltk
numpy
pandas
pymagnitude
pyspellchecker
scikit-learn~=0.24
seaborn
spacy>=3.0
tensorflow
torch
tqdm
xgboost
xlrd
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.1.0/de_core_news_sm-3.1.0-py3-none-any.whl
spacy-sentiws @ git+https://github.com/Liebeck/spacy-sentiws.git@9c05d30d3f206f993c64035c937b7a810698b90b
bokeh
ipywidgets
pyLDAvis
lz4
xxhash
fasteners
hyperopt
lxml
otree==5.10.3
schwifty==2023.1.1
jinja2==2.11.3
MarkupSafe==1.1.1